[add] BI Table abstract class mixin [remove] old BI Table classes
