+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "mobx-lark",
-    "version": "1.0.0-alpha.1",
+    "version": "1.0.0-alpha.3",
     "license": "LGPL-3.0",
     "author": "shiy2008@gmail.com",
     "description": "Unofficial TypeScript SDK for FeiShu/Lark API, which is based on MobX-RESTful.",
