+++ src/index.ts
@@ -5,4 +5,4 @@ globalThis.fetch ||= fetch;
 export * from './type';
 export * from './Lark';
 // export * from './module';
-export * from './model/BiDataTable';
+export * from './module/BITable';
